refactor: decouple layout logic from shared app orchestrator
- Add NeedsCompactAutoFill to the Layout interface so auto-fill
  decisions are owned by each layout, eliminating 9 MillerLayout type
  assertions from app.go
- Extract identical DelegateUpdate bodies into a single
  delegateScreenUpdate function in layout.go; adding a new screen now
  requires one edit instead of two
- Define CompactListRenderer interface (IsCompactListActive, ListTitle,
  CompactListView, DetailView) and implement it on FeedModel, GuildsModel,
  and TopicsModel; MillerLayout.View collapses three separate if-else
  blocks into a single activeCompactRenderer helper
- Move menuTabs, hint, sbStyle, renderHints, overlayCenter, and
  navigateTabBy from layout_tabs.go to layout.go where they belong as
  shared utilities
